Fixing documentation of themes
The CSS paths into package assets have changed with the new Meteor version, starting with `/packages/*user*_*package name*/`

For theme switching the CSS must be load in the HTML by a `link` element with `id="theme"`.

When used with templates the initialization must be done using the `onRendered()` method of the template.